+++ R/TaskVimpHyperparameters.R
@@ -374,6 +374,7 @@ setMethod(
       "settings" = settings,
       "metric" = settings$hpo$hpo_metric,
       "hyperparameters" = settings$vimp$param,
+      "allow_naive_models" = settings$hpo$hpo_allow_naive_models,
       "optimisation_function" = settings$hpo$hpo_optimisation_function,
       "acquisition_function" = settings$hpo$hpo_acquisition_function,
       "grid_initialisation_method" = settings$hpo$hpo_grid_initialisation_method,